Edit Profile Security
Previously, users could change the url of the Edit Profile page to make changes to any profile. This fixes that, by redirecting users back to their own profile page if they attempt to access an Edit Profile page for an account that is not their own. I cannot remember why this change was necessary, I thought it worked well before, but error message after error message kept on hitting me, and this finally solved it (I hope)
idk, its 1:38AM and im tired